Extract environment methods and test them
