+++ rebench/persistence.py
@@ -18,7 +18,6 @@
 # FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
 # IN THE SOFTWARE.
 import json
-import getpass
 import os
 import shutil
 import subprocess
@@ -28,10 +27,8 @@ from tempfile import NamedTemporaryFile
 from threading import Lock
 from time import time
 
-from cpuinfo import get_cpu_info
-from psutil import virtual_memory
-
 from .configuration_error import ConfigurationError
+from .environment import determine_environment, determine_source_details
 from .model.data_point  import DataPoint
 from .model.measurement import Measurement
 from .model.run_id      import RunId
@@ -40,12 +37,10 @@ from .model.run_id      import RunId
 try:
     from http.client import HTTPException
     from urllib.request import urlopen, Request as PutRequest
-    from urllib.parse import urlparse
 except ImportError:
     # Python 2.7
     from httplib import HTTPException
     from urllib2 import urlopen, Request
-    from urlparse import urlparse
 
 
     class PutRequest(Request):
@@ -142,61 +137,8 @@ class DataStore(object):
         return cfg
 
 
-def _encode_str(out):
-    as_string = out.decode('utf-8')
-    if as_string[-1] == '\n':
-        as_string = as_string[:-1]
-    return as_string
-
-
-def _exec(cmd):
-    out = subprocess.check_output(cmd)
-    return _encode_str(out)
-
-
-def _determine_source_details():
-    result = dict()
-    repo_url = subprocess.check_output(['git', 'ls-remote', '--get-url'])
-
-    parsed = urlparse(repo_url)
-    if parsed.password:
-        # remove password
-        parsed = parsed._replace(
-            netloc="{}@{}".format(parsed.username, parsed.hostname))
-    result['repoURL'] = _encode_str(parsed.geturl())
-
-    result['branchOrTag'] = _exec(['git', 'show', '-s', '--format=%D', 'HEAD'])
-    result['commitId'] = _exec(['git', 'rev-parse', 'HEAD'])
-    result['commitMsg'] = _exec(['git', 'show', '-s', '--format=%B', 'HEAD'])
-    result['authorName'] = _exec(['git', 'show', '-s', '--format=%aN', 'HEAD'])
-    result['committerName'] = _exec(['git', 'show', '-s', '--format=%cN', 'HEAD'])
-    result['authorEmail'] = _exec(['git', 'show', '-s', '--format=%aE', 'HEAD'])
-    result['committerEmail'] = _exec(['git', 'show', '-s', '--format=%cE', 'HEAD'])
-    return result
-
-
-def _determine_environment():
-    result = dict()
-    result['userName'] = getpass.getuser()
-    result['manualRun'] = not ('CI' in os.environ and os.environ['CI'] == 'true')
-
-    u_name = os.uname()
-    result['hostName'] = u_name[1]
-    result['osType'] = u_name[0]
-    cpu_info = get_cpu_info()
-    result['cpu'] = cpu_info['brand']
-    result['clockSpeed'] = (cpu_info['hz_advertised_raw'][0]
-                            * (10 ** cpu_info['hz_advertised_raw'][1]))
-    result['memory'] = virtual_memory().total
-    result['software'] = []
-    result['software'].append({'name': 'kernel', 'version': u_name[3]})
-    result['software'].append({'name': 'kernel-release', 'version': u_name[2]})
-    result['software'].append({'name': 'architecture', 'version': u_name[4]})
-    return result
-
-
-_environment = _determine_environment()
-_source = _determine_source_details()
+_environment = determine_environment()
+_source = determine_source_details()
 
 
 class _AbstractPersistence(object):
